requirements.txt
aiofiles
aiohttp
aiohttp_session
cryptography
git+https://github.com/5hojib/pyrogram.git@dev
python-dotenv
tgcrypto
uvloop
